WIP move nonbuffered console out of debugger, add Windows support (untested)
Move the NonBufferedConsole class out of debugger.py and into
lowlevel.py. Dispatch on sys.platform, support POSIX type Unix systems
(select, termios) as well as Windows (msvcrt). Fall back to input()
calls on uncertain platforms (webassembly).

This is mostly untested, but provides a foundation to support all
platforms that the PuDB project is targetting, and remains transparent
to call sites.
